changing yarn values from int to long

+++ src/Core/Models/AmbariResponseEntities/Cluster/AlertsSummary.cs
@@ -10,18 +10,18 @@ namespace Core.Models.AmbariResponseEntities.Cluster
     public class AlertsSummary
     {
         [JsonProperty(PropertyName = "CRITICAL")]
-        public int Critical { get; set; }
+        public long Critical { get; set; }
 
         [JsonProperty(PropertyName = "MAINTENANCE")]
-        public int Maintenance { get; set; }
+        public long Maintenance { get; set; }
 
         [JsonProperty(PropertyName = "OK")]
-        public int Ok { get; set; }
+        public long Ok { get; set; }
 
         [JsonProperty(PropertyName = "UNKNOWN")]
-        public int Unknown { get; set; }
+        public long Unknown { get; set; }
 
         [JsonProperty(PropertyName = "WARNING")]
-        public int Warning { get; set; }
+        public long Warning { get; set; }
     }
 }
